refactor: chat refactoring
