+++ src/components/class/Chat/Notice.jsx
